+++ src/imio/dashboard/events.py
@@ -10,11 +10,16 @@
 from imio.helpers.cache import invalidate_cachekey_volatile_for
 
 
+def onDashboardCollectionCreated(obj, event):
+    '''Called whenever a DashboardCollection is created.'''
+    invalidate_cachekey_volatile_for('imio.dashboard.cachedcollectionvocabulary')
+
+
 def onDashboardCollectionModified(obj, event):
     '''Called whenever a DashboardCollection is modified.'''
-    invalidate_cachekey_volatile_for('imio.dashboard.conditionawarecollectionvocabulary')
+    invalidate_cachekey_volatile_for('imio.dashboard.cachedcollectionvocabulary')
 
 
-def onDashboardCollectionTransition(obj, event):
+def onDashboardCollectionRemoved(obj, event):
     '''Called whenever a WF transition was triggered on a DashboardCollection.'''
-    invalidate_cachekey_volatile_for('imio.dashboard.conditionawarecollectionvocabulary')
+    invalidate_cachekey_volatile_for('imio.dashboard.cachedcollectionvocabulary')