+++ src/imio/dashboard/setuphandlers.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from collective.behavior.talcondition.utils import applyExtender
-
 
 def isNotCurrentProfile(context):
     return context.readDataFile("imiodashboard_marker.txt") is None
@@ -12,6 +10,3 @@ def post_install(context):
     if isNotCurrentProfile(context):
         return
 
-    # we need to apply the ITALConditionable extender on existing elements
-    portal = context.getSite()
-    applyExtender(portal, ('DashboardCollection', ))